[trading-native][storage-core] Extract ShardedJmtMerkleDb substrate from state_merkle_db
Pull the 16-shard + metadata DB + version/LRU cache + JMT/commit/read
infrastructure out of StateMerkleDb into a new ShardedJmtMerkleDb
substrate. StateMerkleDb keeps its is_hot flag and hot/cold path
resolution, opens its DBs the same way as before, and Deref's to
ShardedJmtMerkleDb so all the JMT/commit/read methods keep working on
&StateMerkleDb. TreeReader and TreeWriter are delegated.

Pure refactor on the state side. No observable behavior change in
state's open / commit / read paths. Sets up a single sharded merkle DB
substrate that the position-side wrapper can share in the next commit.
